+++ server/src/api/routes/ShortURL.routes.ts
@@ -1,11 +1,8 @@
 import { Router } from "express";
-import TestApi from "../controllers/Test.controller";
 import CreateShortUrl from "../controllers/ShortURL/CreateShortUrl.controller";
 import RedirectShortUrl from "../controllers/ShortURL/RedirectShortUrl.controller";
 
 const ShortUrlRoutes: Router = Router();
-
-ShortUrlRoutes.get("/test", TestApi);
 
 ShortUrlRoutes.post("/create", CreateShortUrl);